add GitHub workflow that sets up a React Native project with the library
The UITests have been updated to assert ads are rendered. Something similar has to be done for Android implementation.
